Capture chapter headers for display
While analyzing the spacing pattern, capture chapter headings
(following 4 blank lines) into a list, then output as a section
following the spacing pattern output. This ports a feature from
pptools (and improves on it).

Similar to the outline feature in pphtml, but for text files.

This code ignores 2 blanks and only looks for 4 blanks. It makes
no attempt to outline h3/sections, in other words.