Add duplicated lines check
Finds and displays blocks of duplicated lines. A single duplicate
line is ignored, but if 2 or more consecutive lines are seen again
later in the file, they're flagged for display.

Honors verbose mode: in verbose mode, all lines are shown. Otherwise
only 4 lines are shown, plus an indicator more lines matched after.

Blank lines and lines consisting of an asterisk thought-break
pattern are ignored.

This is a feature ported from PPTools "text checks" and is part of a
larger effort to migrate some PPTools features into the PPWB suite.